+++ src/core/as/index.js
@@ -33,7 +33,7 @@ import * as tendermintNdid from '../../tendermint/ndid';
 import * as utils from '../../utils';
 import * as config from '../../config';
 import * as common from '../common';
-import * as db from '../../db';
+import * as cacheDb from '../../db/cache';
 import errorType from '../../error/type';
 
 export * from './register_or_update_as_service';
@@ -148,7 +148,7 @@ export async function afterGotDataFromCallback(
     if (response.status !== 200) {
       const dataRequestId =
         additionalData.requestId + ':' + additionalData.serviceId;
-      db.removeRpIdFromDataRequestId(dataRequestId);
+      cacheDb.removeRpIdFromDataRequestId(dataRequestId);
       throw new CustomError({
         message: errorType.INVALID_HTTP_RESPONSE_STATUS_CODE.message,
         code: errorType.INVALID_HTTP_RESPONSE_STATUS_CODE.code,
@@ -226,7 +226,7 @@ async function getDataAndSendBackToRP(request, responseDetails) {
   await Promise.all(
     request.service_data_request_list.map(async (serviceData) => {
       let { service_id, request_params } = serviceData;
-      const callbackUrl = await db.getServiceCallbackUrl(service_id);
+      const callbackUrl = await cacheDb.getServiceCallbackUrl(service_id);
 
       if (!callbackUrl) {
         logger.error({
@@ -237,7 +237,7 @@ async function getDataAndSendBackToRP(request, responseDetails) {
       }
 
       const dataRequestId = request.request_id + ':' + service_id;
-      await db.setRpIdFromDataRequestId(dataRequestId, request.rp_id);
+      await cacheDb.setRpIdFromDataRequestId(dataRequestId, request.rp_id);
 
       logger.info({
         message: 'Sending callback to AS',
@@ -348,7 +348,7 @@ export async function getServiceDetail(service_id) {
     });
     if (service == null) return null;
     return {
-      url: await db.getServiceCallbackUrl(service_id),
+      url: await cacheDb.getServiceCallbackUrl(service_id),
       min_ial: service.min_ial,
       min_aal: service.min_aal,
       active: service.active,
@@ -364,7 +364,7 @@ export async function getServiceDetail(service_id) {
 
 async function isIdpResponsesValid(request_id, dataFromMq) {
   if (!dataFromMq) {
-    dataFromMq = await db.getRequestReceivedFromMQ(request_id);
+    dataFromMq = await cacheDb.getRequestReceivedFromMQ(request_id);
   }
 
   const {