+++ src/core/idp/index.js
@@ -35,7 +35,7 @@ import * as tendermintNdid from '../../tendermint/ndid';
 import * as common from '../common';
 import * as utils from '../../utils';
 import * as config from '../../config';
-import * as db from '../../db';
+import * as cacheDb from '../../db/cache';
 import * as identity from '../identity';
 
 export * from './create_response';
@@ -299,7 +299,7 @@ export async function processMessage(message) {
         },
       });
     }
-    await db.setRequestMessage(message.request_id, {
+    await cacheDb.setRequestMessage(message.request_id, {
       request_message: message.request_message,
       request_message_salt: message.request_message_salt,
     });
@@ -328,8 +328,8 @@ export async function processIdpResponseAfterAddAccessor(
   try {
     if (error) throw error;
 
-    const reference_id = await db.getReferenceIdByRequestId(message.request_id);
-    const callbackUrl = await db.getCallbackUrlByReferenceId(reference_id);
+    const reference_id = await cacheDb.getReferenceIdByRequestId(message.request_id);
+    const callbackUrl = await cacheDb.getCallbackUrlByReferenceId(reference_id);
     const notifyData = {
       success: true,
       reference_id,
@@ -349,7 +349,7 @@ export async function processIdpResponseAfterAddAccessor(
           },
           true
         );
-        db.removeCallbackUrlByReferenceId(reference_id);
+        cacheDb.removeCallbackUrlByReferenceId(reference_id);
       }
     } else {
       if (callbackUrl == null) {
@@ -364,10 +364,10 @@ export async function processIdpResponseAfterAddAccessor(
           },
           true
         );
-        db.removeCallbackUrlByReferenceId(reference_id);
+        cacheDb.removeCallbackUrlByReferenceId(reference_id);
       }
     }
-    db.removeReferenceIdByRequestId(message.request_id);
+    cacheDb.removeReferenceIdByRequestId(message.request_id);
     await common.closeRequest(
       {
         request_id: message.request_id,
@@ -426,12 +426,12 @@ async function checkCreateIdentityResponse(message) {
     });
     return true;
   } catch (error) {
-    const { associated } = await db.getIdentityFromRequestId(
+    const { associated } = await cacheDb.getIdentityFromRequestId(
       message.request_id
     );
 
-    const reference_id = await db.getReferenceIdByRequestId(message.request_id);
-    const callbackUrl = await db.getCallbackUrlByReferenceId(reference_id);
+    const reference_id = await cacheDb.getReferenceIdByRequestId(message.request_id);
+    const callbackUrl = await cacheDb.getCallbackUrlByReferenceId(reference_id);
     if (associated) {
       if (callbackUrl == null) {
         // Implies API v1
@@ -452,7 +452,7 @@ async function checkCreateIdentityResponse(message) {
           },
           true
         );
-        db.removeCallbackUrlByReferenceId(reference_id);
+        cacheDb.removeCallbackUrlByReferenceId(reference_id);
       }
     } else {
       if (callbackUrl == null) {
@@ -474,10 +474,10 @@ async function checkCreateIdentityResponse(message) {
           },
           true
         );
-        db.removeCallbackUrlByReferenceId(reference_id);
+        cacheDb.removeCallbackUrlByReferenceId(reference_id);
       }
     }
-    db.removeCreateIdentityDataByReferenceId(reference_id);
+    cacheDb.removeCreateIdentityDataByReferenceId(reference_id);
     await common.closeRequest(
       {
         request_id: message.request_id,