+++ src/db/cache/sequelize.js
@@ -23,11 +23,11 @@
 import path from 'path';
 import Sequelize from 'sequelize';
 
-import * as config from '../config';
+import * as config from '../../config';
 
 const dbPath = path.join(
   config.dataDirectoryPath,
-  `db-api-${config.nodeId}.sqlite`
+  `db-cache-api-${config.nodeId}.sqlite`
 );
 
 const sequelize = new Sequelize('ndid-api', null, null, {
@@ -38,7 +38,7 @@ const sequelize = new Sequelize('ndid-api', null, null, {
 });
 
 // Models
-const Entities = {
+export const Entities = {
   expectedTx: sequelize.define('expectedTx', {
     tx: { type: Sequelize.STRING, primaryKey: true },
     metadata: Sequelize.JSON,
@@ -157,135 +157,8 @@ const Entities = {
   })
 };
 
-const initDb = sequelize.sync();
+export const init = sequelize.sync();
 
 export function close() {
   return sequelize.close();
-}
-
-export async function getList({ name, keyName, key, valueName }) {
-  await initDb;
-  const models = await Entities[name].findAll({
-    attributes: [valueName],
-    where: {
-      [keyName]: key,
-    },
-  });
-  return models.map((model) => model.get(valueName));
-}
-
-export async function count({ name, keyName, key }) {
-  await initDb;
-  const count = await Entities[name].count({
-    where: {
-      [keyName]: key,
-    },
-  });
-  return count;
-}
-
-export async function getListRange({ name, keyName, keyRange, valueName }) {
-  await initDb;
-  const models = await Entities[name].findAll({
-    attributes: [valueName],
-    where: {
-      [keyName]: {
-        [Sequelize.Op.gte]: keyRange.gte,
-        [Sequelize.Op.lte]: keyRange.lte,
-      },
-    },
-  });
-  return models.map((model) => model.get(valueName));
-}
-
-export async function pushToList({ name, keyName, key, valueName, value }) {
-  await initDb;
-  await Entities[name].create({
-    [keyName]: key,
-    [valueName]: value,
-  });
-}
-
-export async function removeFromList({
-  name,
-  keyName,
-  key,
-  valueName,
-  valuesToRemove,
-}) {
-  await initDb;
-  await Entities[name].destroy({
-    where: {
-      [keyName]: key,
-      [valueName]: {
-        [Sequelize.Op.in]: valuesToRemove,
-      },
-    },
-  });
-}
-
-export async function removeList({ name, keyName, key }) {
-  await initDb;
-  await Entities[name].destroy({
-    where: {
-      [keyName]: key,
-    },
-  });
-}
-
-export async function removeListRange({ name, keyName, keyRange }) {
-  await initDb;
-  await Entities[name].destroy({
-    where: {
-      [keyName]: {
-        [Sequelize.Op.gte]: keyRange.gte,
-        [Sequelize.Op.lte]: keyRange.lte,
-      },
-    },
-  });
-}
-
-export async function removeAllLists({ name }) {
-  await initDb;
-  await Entities[name].destroy({
-    where: {},
-  });
-}
-
-export async function get({ name, keyName, key, valueName }) {
-  await initDb;
-  const model = await Entities[name].findOne({
-    attributes: [valueName],
-    where: {
-      [keyName]: key,
-    },
-  });
-  return model != null ? model.get(valueName) : null;
-}
-
-export async function set({ name, keyName, key, valueName, value }) {
-  await initDb;
-  await Entities[name].upsert({
-    [keyName]: key,
-    [valueName]: value,
-  });
-}
-
-export async function remove({ name, keyName, key }) {
-  await initDb;
-  await Entities[name].destroy({
-    where: {
-      [keyName]: key,
-    },
-  });
-}
-
-export async function getAll({ name }) {
-  await initDb;
-  const models = await Entities[name].findAll({
-    attributes: {
-      exclude: ['id', 'createdAt', 'updatedAt'],
-    },
-  });
-  return models.map((model) => model.get({ plain: true }));
 }